ai_services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.37.0
torch==2.1.2
beautifulsoup4==4.12.3
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
